+++ src/model/TexImage.cpp
@@ -53,8 +53,8 @@ void TexImage::setHeight(double height) {
     this->calcSize();
 }
 
-auto TexImage::cairoReadFunction(TexImage* image, unsigned char* data, uint64_t length) -> cairo_status_t {
-    for (uint64_t i = 0; i < length; i++, image->read++) {
+auto TexImage::cairoReadFunction(TexImage* image, unsigned char* data, unsigned int length) -> cairo_status_t {
+    for (unsigned int i = 0; i < length; i++, image->read++) {
         if (image->read >= image->binaryData.length()) {
             return CAIRO_STATUS_READ_ERROR;
         }
